+++ certbot-dns-ovh/certbot_dns_ovh/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_ovh.dns_ovh` plugin.
