+++ certbot-dns-ovh/docs/api/dns_ovh.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_dns_ovh.dns_ovh`
+:mod:`certbot_dns_ovh._internal.dns_ovh`
 ------------------------------
 
-.. automodule:: certbot_dns_ovh.dns_ovh
+.. automodule:: certbot_dns_ovh._internal.dns_ovh
    :members:
